Consolidate duplicate alerts; first, on the client, in 20s blocks, then, on the server in 1hr blocks. Display alerts consolidated with date/time range and count on alerts page. Add forced DLL loading for same-arch processes where LoadAppInitDlls() is available. (could be improved)

+++ client/apihook/signatures.h
@@ -77,7 +77,7 @@ inline PHOOKAPI_ACTION_CONF nextActionConf(PHOOKAPI_ACTION_CONF conf){
 //Actions
 enum HookAction { ALERT, BLOCK, KILLPROC, KILLTHREAD };
 
-#define START_INFO (size_t)-1
+#define START_INFO (DWORD)-1
 
 typedef struct sHOOKAPI_FUNC_CONF
 {
